requirements.txt
asgiref==3.2.7
astroid==2.3.3
colorama==0.4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-datatables-view==1.19.1
django-filter==2.2.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.1
psycopg2==2.8.5
pylint==2.4.4
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.0.1
wrapt==1.11.2
